e2e(ssr): renamed to mock-data
