+++ packages/nextjs/mocks/msw/db/mock-data.ts
@@ -5,6 +5,10 @@ type MockData = ReturnType<typeof generateMockData>;
 
 let mockData: null | MockData;
 
+/**
+ * Retrieves the mock data.
+ * Seeds if necessary.
+ */
 export function getMockData() {
   if (!mockData) {
     mockData = generateMockData();
@@ -12,13 +16,18 @@ export function getMockData() {
   return mockData;
 }
 
-export function seedMockData(newMockData: Partial<MockData>) {
-  mockData = Object.assign(generateMockData(), newMockData);
+/**
+ * Sets the mock data
+ */
+export function setMockData(newMockData: Partial<MockData>) {
+  Object.assign(getMockData(), newMockData);
 }
 
-function generateMockData() {
+/**
+ * Generate some random mock data
+ */
+export function generateMockData() {
   const categories = mock.array(5, () => mock.category({}));
-
   const products = mock.array(40, () =>
     mock.product({
       categories: faker.random.arrayElements(categories, faker.datatype.number({ min: 1, max: 3 })),
